requirements.txt
pytest==5.2.1
matplotlib==3.1.1
scipy==1.3.1
seaborn==0.9.0
tqdm==4.35.0
torchvision==0.4.0
pyro_ppl==0.4.1
visdom==0.1.8.9
pandas==0.25.1
torch==1.2.0
numpy==1.17.1
Pillow==6.2.0
scikit_learn==0.21.3
